Move create billing batch event service
