+++ app/services/billing/supplementary/initiate-billing-batch.service.js
@@ -9,7 +9,7 @@ const BillingBatchModel = require('../../../models/water/billing-batch.model.js'
 const ChargingModuleCreateBillRunService = require('../../charging-module/create-bill-run.service.js')
 const CheckLiveBillRunService = require('./check-live-bill-run.service.js')
 const CreateBillingBatchService = require('./create-billing-batch.service.js')
-const CreateBillingBatchEventService = require('./create-billing-batch-event.service.js')
+const CreateBillingBatchEventService = require('../create-billing-batch-event.service.js')
 const ExpandedError = require('../../../errors/expanded.error.js')
 
 /**
