+++ test/services/billing/create-billing-batch-event.service.test.js
@@ -9,14 +9,14 @@ const { describe, it, beforeEach, afterEach } = exports.lab = Lab.script()
 const { expect } = Code
 
 // Test helpers
-const BillingBatchHelper = require('../../../support/helpers/water/billing-batch.helper.js')
-const BillingBatchModel = require('../../../../app/models/water/billing-batch.model.js')
-const DatabaseHelper = require('../../../support/helpers/database.helper.js')
-const EventModel = require('../../../../app/models/water/event.model.js')
-const RegionHelper = require('../../../support/helpers/water/region.helper.js')
+const BillingBatchHelper = require('../../support/helpers/water/billing-batch.helper.js')
+const BillingBatchModel = require('../../../app/models/water/billing-batch.model.js')
+const DatabaseHelper = require('../../support/helpers/database.helper.js')
+const EventModel = require('../../../app/models/water/event.model.js')
+const RegionHelper = require('../../support/helpers/water/region.helper.js')
 
 // Thing under test
-const CreateBillingBatchEventService = require('../../../../app/services/billing/supplementary/create-billing-batch-event.service.js')
+const CreateBillingBatchEventService = require('../../../app/services/billing/create-billing-batch-event.service.js')
 
 describe('Create Event service', () => {
   let clock
